pyproject.toml
[project]
name = "ai-trpg"
version = "0.1.0"
description = "AI-powered Call of Cthulhu TRPG system"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.27.0",
    "socksio>=1.0.0",
]

[project.optional-dependencies]
ai = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
]
db = [
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
